fix: address playlists by usize so large libraries work
Playlist, stream and track indices were u8 throughout the player. The
web UI addresses a playlist by its real position in the library, so
`index as u8` in the route handlers silently wrapped: clicking playlist
398 sent index 142 to the player, which either started the wrong album
or, when the truncated length happened to be smaller, fell through to
the error sound. The MIDI path truncated the library counts the same
way and could overflow when adding them.

Widen the indices to usize and resolve a MIDI source index in a
separate `resolve_source` function so the mapping is testable without
an mpv process.

Also bounds-check the track index in play_playlist_track. It went
straight to `.unwrap()`, so a stale index from the web UI panicked the
process instead of playing the error sound.

+++ src/miconau/web.rs
@@ -209,29 +209,29 @@ async fn get_state(
 }
 
 async fn play_stream(
-    Path(index): Path<u64>,
+    Path(index): Path<usize>,
     State(server_state): State<ServerState>,
 ) -> Result<StatusCode, StatusCode> {
     let mut player = server_state.player.lock().await;
-    player.play_stream(index as u8);
+    player.play_stream(index);
     Ok(StatusCode::OK)
 }
 
 async fn play_playlist(
     State(server_state): State<ServerState>,
-    Path(index): Path<u64>,
+    Path(index): Path<usize>,
 ) -> Result<StatusCode, StatusCode> {
     let mut player = server_state.player.lock().await;
-    player.play_playlist(index as u8);
+    player.play_playlist(index);
     Ok(StatusCode::OK)
 }
 
 async fn play_playlist_track(
     State(server_state): State<ServerState>,
-    Path((index, track_index)): Path<(u64, u64)>,
+    Path((index, track_index)): Path<(usize, usize)>,
 ) -> Result<StatusCode, StatusCode> {
     let mut player = server_state.player.lock().await;
-    player.play_playlist_track(index as u8, track_index as u8);
+    player.play_playlist_track(index, track_index);
     Ok(StatusCode::OK)
 }
